backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
supabase==2.0.0
python-dotenv==1.0.0
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6